Remove positional arguments
Was trying to utilize this module when I kept running into a issue with the returned values for dates just returning "2". By using show-object I was able to determine the REST call was getting a good response back. After seeing that the dates were actually being returned I realized the "2" that was coming back at the end was due to the positional argument being used. 

Tested by removing the [0] for shipDate, warstart, and wed on lines 167-169.